requirements.txt
pygame>=2.5.2
numpy>=1.26
PyYAML>=6.0.1
